Generate density-based clusters.

This function identifies similarly-behaving genomic compartments, such as regions with the highest loss or gain of a specific histone mark.

density_based_scatterplot(
  out_dir,
  genome_assembly,
  treated_samp_norm_bw,
  wildtype_samp_norm_bw,
  are_R_objects = FALSE,
  cell_line,
  histone_mark,
  annotated_clusters,
  number_of_clusters,
  output_filename,
  title_of_plot,
  plot_title_font_size = 12,
  pow = NULL,
  show_legend = FALSE,
  legend_font_size = 7,
  min,
  max,
  bin_size,
  show_scales = TRUE,
  xaxis_label,
  yaxis_label,
  axis_title_font_size = 10,
  height_of_figure = 6,
  width_of_figure = 15
)

Arguments

out_dir

Output directory for the scatterplot.

genome_assembly

Must of be one of hg38 or mm10.

treated_samp_norm_bw

The normalized bigwig file for the treated sample.

wildtype_samp_norm_bw

The normalized bigwig file for the wildtype sample.

are_R_objects

Boolean term (true or false) to indicate if the inputted bigwig files are R objects. It'll use load() for the reps as opposed to reading them in via rtracklayer::import.bed(). Default to FALSE.

cell_line

The cell line of the samples.

histone_mark

The broad histone mark used for the analysis.

annotated_clusters

The annotated clusters generated using annotate_clust(), which is an R object that needs to be loaded.

number_of_clusters

The total number of clusters identified by the HDBSCAN algorithm.

output_filename

Filename for the resuting scatterplot.

title_of_plot

The title of the plot.

plot_title_font_size

An integer specifying the font size for each plot title.

pow

The power of. Returns the value of x to the power of y (x^y) and this is used for the scales (i.e. show bins if surpassing a certain intensity). Defaults to 1.25.

show_legend

Boolean term whether to show legend or not.

legend_font_size

An integer specifying the font size for the legends.

min

The min for the x and y axis.

max

The max for the x and y axis

bin_size

The number of bins represented by each hexagonal shape.

show_scales

Boolean term whether to show scales or not.

xaxis_label

Label for the x-axis. This is normally the baseline label (i.e. WT)

yaxis_label

Label for the y-axis. This is normally the treated sample label.

axis_title_font_size

An integer specifying the font size for the axis titles.

height_of_figure

A numeric specifying the height of the plots.

width_of_figure

A numeric specifying the width of the plots.

Value

a plot of density-based clusters.

Details

Title

Examples

library(tidyverse)
library(rtracklayer)
library(GenomicRanges)
library(data.table)
library(lattice)
library(gridExtra)
library(DiffBind)
library(patchwork)
library(sf)
#> Linking to GEOS 3.12.0, GDAL 3.8.0, PROJ 9.3.0; sf_use_s2() is TRUE
library(MASS)
#> 
#> Attaching package: ‘MASS’
#> The following object is masked from ‘package:patchwork’:
#> 
#>     area
#> The following object is masked from ‘package:dplyr’:
#> 
#>     select
library(lwgeom)
#> Linking to liblwgeom 3.0.0beta1 r16016, GEOS 3.12.0, PROJ 9.3.1
#> Warning: PROJ versions differ: lwgeom has 9.3.1 sf has 9.3.0
#> 
#> Attaching package: ‘lwgeom’
#> The following object is masked from ‘package:sf’:
#> 
#>     st_perimeter
library(ggrepel)
library(hexbin)
library(ggrastr)
library(viridis)
#> Loading required package: viridisLite
library(pals)
#> 
#> Attaching package: ‘pals’
#> The following objects are masked from ‘package:viridis’:
#> 
#>     cividis, inferno, magma, plasma, turbo, viridis
#> The following objects are masked from ‘package:viridisLite’:
#> 
#>     cividis, inferno, magma, plasma, turbo, viridis
library(isoband)
# using data in /extdata folder
# system.file("extdata", "downsampled.Cal27.WT_input.H3K36me2.10kb.bed.gz", package = "ChIPbinner")
density_based_scatterplot(
  out_dir = system.file("extdata", package = "ChIPbinner"),
  genome_assembly = "hg38",
  are_R_objects = TRUE,
  output_filename = "complete",
  wildtype_samp_norm_bw = system.file("extdata/data","Cal27_WT_H3K36me2_10kb_norm_bw.rda", package = "ChIPbinner"),
  treated_samp_norm_bw = system.file("extdata/data","Cal27_NSD1_KO_H3K36me2_10kb_norm_bw.rda", package = "ChIPbinner"),
  cell_line = "Cal27",
  histone_mark = "H3K36me2",
  annotated_clusters = system.file("extdata/data","complete_annotated_clusters.rda", package = "ChIPbinner"),
  number_of_clusters = 3,
  title_of_plot = "H3K36me2",
  pow = 1.1,
  show_legend = TRUE,
  min = -5,
  max = 2,
  bin_size = 50,
  show_scales = FALSE,
  xaxis_label = "WT",
  yaxis_label = "NSD1_KO",
  height_of_figure = 6,
  width_of_figure = 15
)
#> [1] "Density-based scatterplots generated!"
# using data in /data folder
# density_based_scatterplot(
#   out_dir = system.file("extdata", package = "ChIPbinner"),
#   genome_assembly = "hg38",
#   are_R_objects = TRUE,
#   output_filename = "complete",
#   wildtype_samp_norm_bw = system.file("data/Cal27_WT_H3K36me2_10kb_norm_bw.rda", package = "ChIPbinner"),
#   treated_samp_norm_bw = system.file("data/Cal27_NSD1_KO_H3K36me2_10kb_norm_bw.rda", package = "ChIPbinner"),
#   cell_line = "Cal27",
#   histone_mark = "H3K36me2",
#   annotated_clusters = system.file("data/complete_annotated_clusters.rda", package = "ChIPbinner"),
#   number_of_clusters = 3,
#   title_of_plot = "H3K36me2",
#   pow = 1.1,
#   show_legend = TRUE,
#   min = -5,
#   max = 2,
#   bin_size = 50,
#   show_scales = FALSE,
#   xaxis_label = "WT",
#   yaxis_label = "NSD1_KO",
#   height_of_figure = 6,
#   width_of_figure = 15
# )