Merging of normalized bigWig files.

Performs merging of normalized bigWig files, which are normally replicates.

merge_norm_bw(are_R_objects = FALSE, rep1, rep2, merged_samples_label, out_dir)

Arguments

are_R_objects

a logical indicating whether the inputs are bigWig files or R objects.

rep1

a character string specifying the first bigWig file or R object to be merged. Usually the first replicate.

rep2

a character string specifying the second bigWig file or R object to be merged. Usually the second replicate.

merged_samples_label

a character string specifying the filename of the merged bigWig file.

out_dir

a character string specifying the output directory for the merged bigWig file.

Value

a single merged bigWig file.

Details

Title

Examples

merge_norm_bw(are_R_objects = TRUE,
              rep1 = system.file("extdata/data","PA_1.rda", package = "ChIPbinner"),
              rep2 = system.file("extdata/data","PA_2.rda", package = "ChIPbinner"),
              merged_samples_label = "merged_PA",
              out_dir = system.file("extdata",package = "ChIPbinner"))