Normalization and conversion of binned BED files into bigWig.

Normalizes genome-wide binned scores and generates a bigWig file containing these normalized scores along with their corresponding genomic regions.

norm_bw(
  out_dir,
  genome_assembly,
  chip_samp_binned_file,
  chip_samp_library_size,
  use_control = TRUE,
  control_binned_file = NULL,
  control_library_size = NULL,
  pseudocount = NULL,
  raw_count_cutoff = NULL,
  scaling_factor = NULL,
  per_million = FALSE
)

Arguments

out_dir

a character string specifying the output directory for the resulting bigWig files.

genome_assembly

a character string specifying the genome assembly. Allowed values include "hg38" or "mm10".

chip_samp_binned_file

a character string specifying the binned BED file for the treated sample.

chip_samp_library_size

an integer specifying the mapped library size for the treated sample.

use_control

a logical indicating whether to use a control sample or not.

control_binned_file

a character string specifying the binned BED file for the control sample. This is usually the genomic input.

control_library_size

an integer specifying the mapped library size for the control sample.

pseudocount

a numeric specifying the pseudocount to avoid division by 0. If NULL (the default), the pseudocount is 1e-3.

raw_count_cutoff

a numeric specifying the raw read count cutoff to exclude bins depleted in signal across all tracks. For example, if 10 is inputted, this removes raw read count consistently lower than 10. If NULL (the default), the raw count cutoff is 0.

scaling_factor

a numeric specifying the quantitative normalization/scaling of the raw binned signal by this factor. For example, this could be ChIP-Rx values or genome-wide modification percentage values obtained from mass spectrometry values. If NULL(the default), no quantitative scaling is performed.

per_million

a logical indicating whether to scale the normalized scores by millions.

Value

a single bigWig file with normalized binned scores.

Details

Title

Examples

norm_bw(out_dir = system.file("extdata",package = "ChIPbinner"),
        genome_assembly = "hg38",
        chip_samp_binned_file = system.file("extdata", "downsampled.Cal27.WT.H3K36me2.10kb.bed.gz", package = "ChIPbinner"),
        chip_samp_library_size = 64093770,
        use_control = TRUE,
        control_binned_file = system.file("extdata", "downsampled.Cal27.WT_input.H3K36me2.10kb.bed.gz", package = "ChIPbinner"),
        control_library_size = 52047022,
        raw_count_cutoff = 0,
        pseudocount = 1e-15,
        scaling_factor = 0.450328805)
#> [1] "ChIP/Cut&Run sample=downsampled.Cal27.WT.H3K36me2.10kb.bed"
#> [1] "ChIP/Cut&Run sample sequencing depth=64093770"
#> [1] "Control sample=downsampled.Cal27.WT_input.H3K36me2.10kb.bed"
#> [1] "Control sample sequencing depth=52047022"
#> [[1]]
#> BigWigFile object
#> resource: /private/var/folders/x6/2pf8nnc52hsfqb4j3ttppj_r0000gn/T/RtmputrD48/temp_libpath1667f6d2c865f/ChIPbinner/extdata/downsampled.Cal27.WT.H3K36me2.10kb.bed.norm.bw 
#>